+++ examples/integrations/mastra/tsconfig.channel.json
@@ -26,5 +26,5 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["channel-host.mts", "src/agent.ts"]
+  "include": ["channel-host.mts", "channels.mts", "src/agent.ts"]
 }
