+++ examples/integrations/mcp-apps/tsconfig.channel.json
@@ -7,5 +7,5 @@
     "skipLibCheck": true,
     "strict": true
   },
-  "include": ["channel-host.mts", "app/agent.ts"]
+  "include": ["channel-host.mts", "channels.mts", "app/agent.ts"]
 }
